fix(F-207): move header and footer to components folder
