+++ frontend/src/app/[locale]/(main)/history/components/HistoryHeader.tsx
