+++ frontend/src/app/[locale]/(main)/new-training/components/CategoryStep.tsx
