+++ frontend/src/app/[locale]/(main)/new-training/components/CustomizeStep.tsx
