+++ frontend/src/app/[locale]/(main)/new-training/components/SituationStep.tsx
